requirements.txt
matplotlib==3.0.2
Click==7.0
numpy==1.15.4
setuptools==40.6.3
tqdm==4.28.1
seaborn==0.9.0
pandas==0.23.4
python-dotenv
torch==1.0.0
appdirs
pyyaml
pytorch-ignite
opencv-python
Pillow
visdom
scikit-learn
scikit-video